+++ gen/client/model_network_identifier.go
@@ -23,6 +23,7 @@ const (
 	NETWORKIDENTIFIER_BASE_SEPOLIA NetworkIdentifier = "base-sepolia"
 	NETWORKIDENTIFIER_BASE_MAINNET NetworkIdentifier = "base-mainnet"
 	NETWORKIDENTIFIER_ETHEREUM_HOLESKY NetworkIdentifier = "ethereum-holesky"
+	NETWORKIDENTIFIER_ETHEREUM_HOODI NetworkIdentifier = "ethereum-hoodi"
 	NETWORKIDENTIFIER_ETHEREUM_SEPOLIA NetworkIdentifier = "ethereum-sepolia"
 	NETWORKIDENTIFIER_ETHEREUM_MAINNET NetworkIdentifier = "ethereum-mainnet"
 	NETWORKIDENTIFIER_POLYGON_MAINNET NetworkIdentifier = "polygon-mainnet"
@@ -40,6 +41,7 @@ var AllowedNetworkIdentifierEnumValues = []NetworkIdentifier{
 	"base-sepolia",
 	"base-mainnet",
 	"ethereum-holesky",
+	"ethereum-hoodi",
 	"ethereum-sepolia",
 	"ethereum-mainnet",
 	"polygon-mainnet",